+++ Zoo_DataBase(Bunina)/Dog.java
@@ -1,19 +1,17 @@
 public class Dog extends Animal {
 
-    public Dog(String name, int age) {
-        this.name = name;
-        this.age = age;
-        System.out.println(name +" "+ age);
+    public Dog(String name, int age,String colour) {
+        super(name,age,colour);
     }
+
     @Override
     public void voice() {
         System.out.println("Гав-гав");
     }
 
-
     @Override
     public void eat() {
         System.out.println("Вы покормили собаку");
-        }
+    }
 
 }
