--- project/documentation/README.md
@@ -1,2 +1,0 @@
-# Infotech2016
-Крутейший проект наикласснейших ребят из учебного центра NetCracker 2016 года!
